Get rid of the states experiment.
