+++ src/session_types/mod.rs
@@ -1,19 +1,79 @@
-/// Inspired by https://github.com/Munksgaard/session-types
-/// See the parent module for more description.
+/// Inspired by:
+/// * https://github.com/Munksgaard/session-types
+/// * https://github.com/input-output-hk/ouroboros-network/#ouroboros-network-documentation
+/// * https://github.com/input-output-hk/ouroboros-network/blob/master/typed-protocols/src/Network/TypedProtocol/Core.hs
+///
+/// The following changes from the original Rust library:
+/// * Send dynamic types in the channel that can be downcast into specific types the protocol expects,
+///   so that we can handle mismatches gracefully, without segfaults, assuming that they are coming over the network
+///   and cannot be trusted to be correct. The compiler is there to help get the protocol right for honest participants,
+///   but in itself cannot guarantee what the right content will arrive.
+/// * Do away without having to send `true` or `false` to indicate choice in `Choose` and `Offer`.
+///   That means `Choose<Send, Recv>` will no longer be possibl, because the other side wouldn't know what we decided;
+///   there has to be an explicit message communicating the choice, and transfering the agency to the opposite peer.
+/// * Returning `Result` type so errors can be inspected. An error closes the channel to be closed.
 use std::{
+    any::Any,
     error::Error,
     marker,
     marker::PhantomData,
     mem::ManuallyDrop,
-    sync::mpsc::{self, Receiver, Sender},
+    sync::mpsc::{self, Receiver, RecvError, RecvTimeoutError, Sender},
     thread,
     time::Duration,
 };
 
-pub use Branch::*;
+type DynMessage = Box<dyn Any + marker::Send + 'static>;
 
-use super::{downcast, DynMessage};
-pub use super::{ok, SessionError, SessionResult};
+#[derive(Debug)]
+pub enum SessionError {
+    /// Wrong message type was sent.
+    UnexpectedMessage(DynMessage),
+    /// The other end of the channel is closed.
+    Disconnected,
+    /// Did not receive a message within the timeout.
+    Timeout,
+    /// Abort due to the a violation of some protocol constraints.
+    Abort(Box<dyn Error + marker::Send + 'static>),
+}
+
+impl From<RecvError> for SessionError {
+    fn from(_: RecvError) -> Self {
+        SessionError::Disconnected
+    }
+}
+
+impl From<RecvTimeoutError> for SessionError {
+    fn from(e: RecvTimeoutError) -> Self {
+        match e {
+            RecvTimeoutError::Disconnected => SessionError::Disconnected,
+            RecvTimeoutError::Timeout => SessionError::Timeout,
+        }
+    }
+}
+
+impl std::fmt::Display for SessionError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
+
+impl Error for SessionError {}
+
+pub type SessionResult<T> = Result<T, SessionError>;
+
+pub fn ok<T>(value: T) -> SessionResult<T> {
+    Ok(value)
+}
+
+fn downcast<T: 'static>(msg: DynMessage) -> SessionResult<T> {
+    match msg.downcast::<T>() {
+        Ok(data) => Ok(*data),
+        Err(invalid) => Err(SessionError::UnexpectedMessage(invalid)),
+    }
+}
+
+pub use Branch::*;
 
 /// A session typed channel. `P` is the protocol and `E` is the environment,
 /// containing potential recursion targets.
@@ -311,11 +371,13 @@ pub fn session_channel<P: HasDual>() -> (Chan<P, ()>, Chan<P::Dual, ()>) {
 ///
 /// ```rust
 /// use paxos::offer;
-/// use paxos::session::types::*;
+/// use paxos::session_types::*;
 /// use std::thread::spawn;
 /// use std::time::Duration;
 ///
-/// fn srv(c: Chan<(), Offer<Recv<u64, Eps>, Offer<Recv<String, Eps>, Eps>>>) -> SessionResult<()>{
+/// struct Bye;
+///
+/// fn srv(c: Chan<Offer<Recv<u64, Eps>, Offer<Recv<String, Eps>, Recv<Bye, Eps>>>, ()>) -> SessionResult<()>{
 ///     let t = Duration::from_secs(1);
 ///     offer! { c, t,
 ///         Number => {
@@ -327,12 +389,12 @@ pub fn session_channel<P: HasDual>() -> (Chan<P, ()>, Chan<P::Dual, ()>) {
 ///             c.recv(t)?.0.close()
 ///         },
 ///         Quit => {
-///             c.close()
+///             c.recv(t)?.0.close()
 ///         }
 ///     }
 /// }
 ///
-/// fn cli(c: Chan<(), Choose<Send<u64, Eps>, Choose<Send<String, Eps>, Eps>>>) -> SessionResult<()>{
+/// fn cli(c: Chan<Choose<Send<u64, Eps>, Choose<Send<String, Eps>, Send<Bye, Eps>>>, ()>) -> SessionResult<()>{
 ///     c.sel1().send(42)?.close()
 /// }
 ///
@@ -348,166 +410,19 @@ pub fn session_channel<P: HasDual>() -> (Chan<P, ()>, Chan<P::Dual, ()>) {
 #[macro_export]
 macro_rules! offer {
     (
-        $id:ident, $timeout:ident, $branch:ident => $code:expr, $($t:tt)+
+        $id:ident, $timeout:expr, $branch:ident => $code:expr, $($t:tt)+
     ) => (
         match $id.offer($timeout)? {
-            $crate::session::types::Left($id) => $code,
-            $crate::session::types::Right($id) => offer!{ $id, $timeout, $($t)+ }
+            $crate::session_types::Left($id) => $code,
+            $crate::session_types::Right($id) => offer!{ $id, $timeout, $($t)+ }
         }
     );
     (
-        $id:ident, $timeout:ident, $branch:ident => $code:expr
+        $id:ident, $timeout:expr, $branch:ident => $code:expr
     ) => (
         $code
     )
 }
 
 #[cfg(test)]
-mod test {
-    use super::*;
-    use std::{thread, time::Instant};
-
-    mod ping_pong {
-        use super::*;
-        pub struct Ping;
-        pub struct Pong;
-
-        // NOTE: This verison doesn't include looping.
-        pub type Server = Recv<Ping, Send<Pong, Eps>>;
-        pub type Client = <Server as HasDual>::Dual;
-    }
-
-    #[test]
-    fn ping_pong_basics() {
-        use ping_pong::*;
-        let t = Duration::from_millis(100);
-
-        let srv = move |c: Chan<Server, ()>| {
-            let (c, _ping) = c.recv(t)?;
-            c.send(Pong)?.close()
-        };
-
-        let cli = move |c: Chan<Client, ()>| {
-            let c = c.send(Ping)?;
-            let (c, _pong) = c.recv(t)?;
-            c.close()
-        };
-
-        let (server_chan, client_chan) = session_channel();
-
-        let srv_t = thread::spawn(move || srv(server_chan));
-        let cli_t = thread::spawn(move || cli(client_chan));
-
-        srv_t.join().unwrap().unwrap();
-        cli_t.join().unwrap().unwrap();
-    }
-
-    #[test]
-    fn ping_pong_error() {
-        use ping_pong::*;
-        let t = Duration::from_secs(10);
-
-        type WrongClient = Send<String, Recv<u64, Eps>>;
-
-        let srv = move |c: Chan<Server, ()>| {
-            let (c, _ping) = c.recv(t)?;
-            c.send(Pong)?.close()
-        };
-
-        let cli = move |c: Chan<WrongClient, ()>| {
-            let c = c.send("Hello".into())?;
-            let (c, _n) = c.recv(t)?;
-            c.close()
-        };
-
-        let (server_chan, client_chan) = session_channel();
-        let wrong_client_chan = client_chan.cast::<WrongClient, ()>();
-
-        let srv_t = thread::spawn(move || srv(server_chan));
-        let cli_t = thread::spawn(move || cli(wrong_client_chan));
-
-        let sr = srv_t.join().unwrap();
-        let cr = cli_t.join().unwrap();
-
-        assert!(sr.is_err());
-        assert!(cr.is_err());
-    }
-
-    #[test]
-    fn greetings() {
-        struct Hail(String);
-        struct Greetings(String);
-        struct TimeRequest;
-        struct TimeResponse(Instant);
-        struct AddRequest(u32);
-        struct AddResponse(u32);
-        struct Quit;
-
-        type TimeProtocol = Recv<TimeRequest, Send<TimeResponse, Var<Z>>>;
-        type AddProtocol = Recv<AddRequest, Recv<AddRequest, Send<AddResponse, Var<Z>>>>;
-        type QuitProtocol = Recv<Quit, Eps>;
-
-        type ProtocolChoices = Offer<TimeProtocol, Offer<AddProtocol, QuitProtocol>>;
-
-        type Server = Recv<Hail, Send<Greetings, Rec<ProtocolChoices>>>;
-        type Client = <Server as HasDual>::Dual;
-
-        // It is at this point that an invalid protocol would fail to compile.
-        let (server_chan, client_chan) = session_channel::<Server>();
-
-        let srv = |c: Chan<Server, ()>| {
-            let t = Duration::from_millis(100);
-            let (c, Hail(cid)) = c.recv(t)?;
-            let c = c.send(Greetings(format!("Hello {}!", cid)))?;
-            let mut c = c.enter();
-            loop {
-                c = offer! { c, t,
-                    Time => {
-                        let (c, TimeRequest) = c.recv(t)?;
-                        let c = c.send(TimeResponse(Instant::now()))?;
-                        c.zero()
-                    },
-                    Add => {
-                        let (c, AddRequest(a)) = c.recv(t)?;
-                        let (c, AddRequest(b)) = c.recv(t)?;
-                        let c = c.send(AddResponse(a + b))?;
-                        c.zero()
-                    },
-                    Quit => {
-                        let (c, Quit) = c.recv(t)?;
-                        c.close()?;
-                        break;
-                    }
-                };
-            }
-
-            ok(())
-        };
-
-        let cli = |c: Chan<Client, ()>| {
-            let t = Duration::from_millis(100);
-            let c = c.send(Hail("Rusty".into()))?;
-            let (c, Greetings(_)) = c.recv(t)?;
-            let c = c.enter();
-            let (c, AddResponse(r)) = c
-                .sel2()
-                .sel1()
-                .send(AddRequest(1))?
-                .send(AddRequest(2))?
-                .recv(t)?;
-
-            c.zero().sel2().sel2().send(Quit)?.close()?;
-
-            ok(r)
-        };
-
-        let srv_t = thread::spawn(move || srv(server_chan));
-        let cli_t = thread::spawn(move || cli(client_chan));
-
-        let sr = srv_t.join().unwrap();
-        let cr = cli_t.join().unwrap();
-
-        assert!(sr.is_ok());
-        assert_eq!(cr.unwrap(), 3);
-    }
-}
+mod test;
